pyproject.toml
[project]
name = "rxoms"
version = "0.1.0"
description = "Runtime eXplainability for Orchestrating ML-base Security"
authors = [
  { name = "AaltoSEA", email = "tri.m.nguyen@aalto.fi" },
  { name = "nguu0123", email = "nguu0123@gmail.com" },
]
license = { text = "Apache-2.0" }
keywords = ["Runtime Explainability"]
homepage = "https://rdsea.github.io/"
repository = "https://github.com/rdsea/RXOMS"
#documentation = "https://rdsea.github.io/QoA4ML/"
readme = "README.md"

dependencies = [
    "pandas",
    "numpy",
    "pydantic==2.7.4",
    "pika==1.3.2",
    "pyyaml",
    "uvicorn",
    "fastapi[standard]",
    "requests",
    "flask>=3.0.3",
    "flask-restful>=0.3.10",
    "setuptools==59.5.0",
    "eventlet==0.30.2",
    "qoa4ml==0.3.16",
    "pymongo==4.10.1",
]
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rxoms"]

[tool.rye]
dev-dependencies = [
  "pre-commit",
  "ruff",
  "codespell",
  # "pip>=24.2",
  # "pytest>=8.3.2",
  # "tox>=4.18.0",
  # "pytest-sugar>=1.0.0",
  # "tox-uv>=1.11.2",
  # "tox-gh-actions>=3.2.0",
]
managed = true
universal = true


[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 10

[tool.ruff.lint.pydocstyle]
convention = "pep257" # Accepts: "google", "numpy", or "pep257".

[tool.ruff.lint]
exclude = ["src/qoa4ml/utils/pynvml_forked.py"]
select = [
  "F",   # Pyflakes
  "E",   # Pycode styles
  "W",   # Pycode styles
  "N",   # Pep8 naming
  "I",   # Isort
  "UP",  # Pyupgrade
  "C4",  # Flake8 comprehension
  "ISC", # Flake8 implicit str concat
  "B",   # Flake8 bugbear
  "PGH", # Pygrep hook
  "RUF", # Ruff specific rules
]
ignore = [
  "E501",   # Line too long
  "RET504", # Unnecessary assignment
  "RET505", # Unnecessary else after return
  "C901",   # Too complex function
  "B904",   # raise-without-from-inside-except, enable when understand more
  "PLR",
  "ISC001",
]

[tool.rye.lint.pylint]
max-args = 10
